Account numbers by account endpoint, error handling
Adds account numbers by account guid endpoint which is part of account
verification. Also adds error handling to `json.Unmarshal` references.
